Add support for PMIX_MEM_ALLOC_KIND
Add a new cmd line option that corresponds to this
attribute. Add the attribute to the prun payload.
When received, it will default to including in the
job info for the spawned job. Add query support for it.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 39577895f6c2beb7531e4a0707ce451b5c78de91)