Fix deprecation warnings for ppr on socket objects
We have deprecated the "socket" object in favor of
"package", so we need to extend that treatment to
the resource qualifier in the `--map-by ppr` directive.
Detect both "socket" and the "skt" shorthand for
backward-compatibility reasons.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(from upstream commit 7f507dfe5fc4e571bc1db8c3efb1db8f521209cd)