Restore parsing of OMPI param files
It isn't perfect and it does raise precedence
concerns, but let the OMPI schizo component
parse the OMPI param files. Use the PMIx
pmdl functions to check for relevant params
to convert to PRRTE and PMIx. Restore the
overlap detection for when params are set
for OMPI frameworks that have a corrolary
to frameworks in PRRTE and PMIx.

This still raises precedence questions. For
now, what this will do is have values in the
OMPI param files give way to corresponding
values specified in PRRTE and PMIx param
files. In other words, values in PMIx and
PRRTE param files are written into the
environment, not overwriting any previously
existing envar. OMPI param files are then
read, and any params that correspond to
PRRTE and PMIx params are written into
the environment - without overwriting
any that already exist.

So if you have a param specified in a
PRRTE param file, and you also specify it
in the OMPI param file, the PRRTE param file
value will "win".

User beware - someone is bound to be very
unhappy.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(from upstream commit fcbefcc9bcbe187ed44795a4453dbb155c8ef352)